UI styles clean up
