+++ package.json
@@ -18,7 +18,6 @@
   "devDependencies": {
     "@codemirror/lang-html": "^6.4.1",
     "@codemirror/state": "^6.2.0",
-    "@codemirror/theme-one-dark": "^6.1.0",
     "@codemirror/view": "^6.7.3",
     "@reduxjs/toolkit": "^1.9.1",
     "@tailwindcss/forms": "^0.5.3",
